add display_board method into tic_tac_toe file
