+++ lib/tic_tac_toe.rb
@@ -58,3 +58,10 @@ def find_winner_in_diagonal(board)
     end
 
 end
+
+def display_board(console, board)
+    board.each do |row|
+        console.puts row.join(" ")
+    end
+end
+
